+++ R/terminal_nodes.R
@@ -16,34 +16,26 @@ terminalNodesDepth = function(model){
   depth     = NULL
   num_trees = model[["num.trees"]]
 
-  # per tree node ids and depth
-  nodeDepth = function(x){
-
-    out = terminalNodesDepthPerTree(ranger::treeInfo(model, x))
-    # return two column df for that tree: id_node and depth
-    res = data.table::data.table(id_node = names(out)
-                                 , depth = unname(out)
-                                 )
-    return(res)
-  }
-
   # iterate over trees
-  tnd = future.apply::future_lapply(1:num_trees, nodeDepth)
+  tnd = future.apply::future_lapply(
+    1:num_trees
+    , function(x) terminalNodesDepthPerTree(ranger::treeInfo(model, x))
+    )
   tnd = data.table::rbindlist(tnd, idcol = "id_tree")
 
   tnd[ , id_node := as.integer(id_node)]
   tnd[ , depth   := as.integer(depth)]
 
-  return(tnd)
+  return(tnd[])
 }
 
 
 #' @name terminalNodesDepthPerTree
 #' @title Depth of each terminal node of a single tree in a ranger model
 #' @description Depth of each terminal node of a single tree in a ranger model.
-#'   Note that root node has the node_id = 0.
+#'   Note that root node has the id_node = 0.
 #' @param treelike Output of `ranger::treeInfo`
-#' @return Named vector of depths where the names are the terminal node IDs
+#' @return data.table with two columns: id_node and depth
 #' @examples
 #' \dontrun{
 #'   rf = ranger::ranger(Species ~ ., data = iris)
@@ -51,28 +43,27 @@ terminalNodesDepth = function(model){
 #' }
 terminalNodesDepthPerTree = function(treelike){
 
-  nodeID     = NULL
-  leftChild  = NULL
-  rightChild = NULL
-  terminal   = NULL
-
   treelike  = data.table::as.data.table(treelike)
   melted    = data.table::melt(
     treelike[, c("nodeID", "leftChild", "rightChild")]
     , id.vars      = "nodeID"
     , measure.vars = c("leftChild", "rightChild")
+    , na.rm        = TRUE
     )
-  value       = NULL
-  edgeMat     = as.matrix(melted[!is.na(value), c("nodeID", "value")]) + 1L
-  treegraph   = igraph::graph_from_edgelist(edgeMat)
-  tnValues    = treelike[terminal == TRUE,][["nodeID"]]
+
+  # create graph in 1-index mode
+  edgeMat     = as.matrix(melted[ , c("nodeID", "value")]) + 1L
+  treegraph   = igraph::graph_from_edgelist(edgeMat, directed = TRUE)
+
+  # get depths of terminal nodes
+  tnValues    = treelike[["nodeID"]][treelike[["terminal"]]]
   depths      = igraph::distances(treegraph
-                                   , v    = 1
-                                   , to   = tnValues + 1L
+                                   , v    = 1L
+                                   , to   = tnValues + 1L # to 1-index mode
                                    , mode = "out"
                                    )
-  dim(depths)   = NULL
-  names(depths) = tnValues
+  dim(depths) = NULL
 
-  return(depths)
+  res = data.table::data.table(id_node = tnValues, depth = depths)
+  return(res)
 }